+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.10] - 2023-08-30
+
+### Fixed
+
+- Improved `VvCombobox` support for object values.
+
 ## [0.0.9] - 2023-08-30
 
 ### Fixed
@@ -24,8 +30,8 @@ All notable changes to this project will be documented in this file.
 - Default props by global plugin configuration for input components;
 - Emit of transition events on `VvDialog` and `VvDropdown`;
 - `useAlert` to manage alerts, story and test;
-- `VvCombobox` custom search function.
-- `VvInputText` add `select-on-focus` prop to select all text on focus;
+- `VvCombobox` custom search function;
+- `VvInputText` add `select-on-focus` prop to select all text on focus.
 
 ## [0.0.8] - 2023-05-22
 
@@ -172,6 +178,7 @@ All notable changes to this project will be documented in this file.
 - `VvTextarea` component;
 - `VvRadioGroup` component.
 
+[0.0.10]: https://github.com/volverjs/style/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/volverjs/style/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/volverjs/style/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/volverjs/style/compare/v0.0.6...v0.0.7